+++ client/src/index.js
@@ -1,9 +1,10 @@
-import firebase from 'firebase'
+import firebase from 'firebase/app'
+import 'firebase/auth'
 import credential from '../credential'
 
 // firebase の初期化
 firebase.initializeApp(credential)
-
+import pb from './word_war_grpc_web_pb'
 // Parcel を使っているので直接 import できる
 const { Elm } = require('./main.elm')
 
@@ -11,10 +12,134 @@ const { Elm } = require('./main.elm')
 const app = Elm.Main.init({
   node: document.getElementById('app'),
   // Elm flags
-  flags: "Hoge"
+  flags: "Hoge",
 })
 
-export default {
-  ports: app.ports,
-  firebase
-}
+const STUBBED = true
+
+const store = {}
+
+// gRPC API のエンドポイント
+const endpoint = 'http://localhost:50051'
+// gRPC のクライアント
+const client = new pb.WordWarPromiseClient()
+
+// ====================
+//        Top
+// ====================
+
+// ログインボタンが押されて Elm から呼ばれる
+app.ports.signinWithFirebase.subscribe(async ({ email, password }) => {
+  // firebase auth でユーザー認証
+  const auth = await firebase.auth()
+    .signInWithEmailAndPassword(email, password)
+    .catch(console.log)
+
+  if (!auth) {
+    return
+  }
+
+  console.log(`logged in as ${auth.user.uid}`)
+
+  app.ports.signinCallback.send({ uid: auth.user.uid })
+})
+
+// サインアップボタンが押されたときの処理
+app.ports.signupWithFirebase.subscribe(async ({ email, password }) => {
+  console.log({ email, password })
+  const auth = await firebase.auth()
+    .createUserWithEmailAndPassword(email, password)
+    .catch(console.error)
+
+  if (!auth) {
+    return
+  }
+
+  console.log(`signed up and logged in as ${auth.user.uid}`)
+
+  app.ports.signinCallback.send({ uid: auth.user.uid })
+})
+
+// ====================
+//        Home
+// ====================
+
+
+// ====================
+//        Game
+// ====================
+app.ports.startGame.subscribe(async userId => {
+  console.log('start game')
+  if (!STUBBED) {
+    // Server Streaming RPCs
+    const req = new pb.GameRequest()
+    req.setUserId(userId)
+    const stream = client.game(req)
+
+    stream.on('data', res => {
+      const message = res.getMessage()
+      const userId = req.getUserId()
+      // JS -> Elm
+      app.ports.onMessage.send({ userId, message })
+    })
+
+    stream.on('end', () => {
+      app.ports.onFinish.send(null)
+    })
+  } else {
+    return new Promise(async () => {
+      // stub
+      store.messages = []
+      for (let i = 0; i < 5; i++) {
+        app.ports.onMessage.send({ userId: 'u012', message: 'message' })
+        await new Promise(resolve => setTimeout(resolve, 2000))
+      }
+
+      app.ports.onFinish.send(null)
+    })
+  }
+})
+
+app.ports.say.subscribe(async ({ userId, message }) => {
+  console.log({ userId, message })
+
+  if (!STUBBED) {
+    // gRPC Unary RPCs
+    const req = new pb.SayRequest()
+    req.setUserId(userId)
+    req.setMessage(message)
+
+    const [ userId, message ] = await client.Say(req)
+      .then(res => res.array)
+      .catch(console.error)
+
+    // Elm の onMessage を呼ぶ
+    app.ports.onMessage.send({ userId, message })
+  } else {
+    store.messages.push({ userId, message })
+    app.ports.onMessage.send({ userId, message })
+  }
+})
+
+// ====================
+//       Result
+// ====================
+
+app.ports.requestResult.subscribe(async userId => {
+  if (!STUBBED) {
+    const req = new pb.ResultRequest()
+    req.setUserId(userId)
+
+    const [ userId, score ] = await client.Result(req)
+      .then(res => res.array)
+      .catch(console.error)
+
+    app.ports.onResult.send({ userId, score })
+  } else {
+    const score = store.messages.length
+    console.log({ userId, score })
+    app.ports.onResult.send({ userId, score })
+  }
+})
+
+export default app
